requirements.txt
numpy==1.16.2
Pillow==9.5.0
PyYAML==6.0.1
Requests==2.31.0
tflite_runtime==2.5.0
